requirements.txt
paho-mqtt
pymongo[srv]
python-dotenv
dnspython
hashlib